Add worker factory
